move exchanges to folder
